Move cookieFunctions content to site.js to as this holds various js scripts related to the website. Added script to summon the print dialogue and this includes to content change in the view for AC4 of the parent ticket 211064
