+++ CheckYourEligibility.Admin/Views/Check/ApplicationsRegistered.cshtml
@@ -20,45 +20,50 @@
             <h3 class="govuk-notification-banner__heading">
                 Children's details added
             </h3>
-            @if( isLA == true)
+            @if (isLA == true)
             {
                 <p class="govuk-body">You've successfully added details to @Model.First().Data.ParentFirstName @Model.First().Data.ParentLastName free school meals application record.</p>
             }
-            else{
-            <p class="govuk-body">You've successfully added details to the free school meals application record
-                for @Model.First().Data.ParentFirstName @Model.First().Data.ParentLastName</p>
+            else
+            {
+                <p class="govuk-body">
+                    You've successfully added details to the free school meals application record
+                    for @Model.First().Data.ParentFirstName @Model.First().Data.ParentLastName
+                </p>
             }
         </div>
     </div>
 
-    <h1 class="govuk-heading-l">Application record
-        for @Model.First().Data.ParentFirstName @Model.First().Data.ParentLastName</h1>
+    <h1 class="govuk-heading-l">
+        Application record
+        for @Model.First().Data.ParentFirstName @Model.First().Data.ParentLastName
+    </h1>
 
     <div class="govuk-grid-row">
         <div class="govuk-grid-column-full">
             <table class="govuk-table">
                 <thead class="govuk-table__head">
-                <tr class="govuk-table__row">
-                    <th class="govuk-table__header">Child</th>
+                    <tr class="govuk-table__row">
+                        <th class="govuk-table__header">Child</th>
                         @if (isLA == true)
                         {
                             <th class="govuk-table__header">School</th>
                         }
-                    <th class="govuk-table__header">Reference</th>
-                </tr>
+                        <th class="govuk-table__header">Reference</th>
+                    </tr>
                 </thead>
                 <tbody class="govuk-table__body">
-                @foreach (var item in Model)
-                {
-                    <tr class="govuk-table__row">
-                        <td class="govuk-table__cell">@item.Data.ChildFirstName @item.Data.ChildLastName</td>
+                    @foreach (var item in Model)
+                    {
+                        <tr class="govuk-table__row">
+                            <td class="govuk-table__cell">@item.Data.ChildFirstName @item.Data.ChildLastName</td>
                             @if (isLA == true)
                             {
                                 <td class="govuk-table__cell">@item.Data.Establishment.Name, @item.Data.Establishment.LocalAuthority.Name </td>
                             }
-                        <td class="govuk-table__cell">@item.Data.Reference</td>
-                    </tr>
-                }
+                            <td class="govuk-table__cell">@item.Data.Reference</td>
+                        </tr>
+                    }
                 </tbody>
             </table>
         </div>
@@ -68,14 +73,25 @@
         <div class="govuk-grid-row">
             <div class="govuk-grid-column-full">
                 <h2 class="govuk-heading-m">What to do next</h2>
-                <p>As these children are entitled to free school meals, you’ll now need to add details of their
-                    application to your own system before finalising.</p>
-                <p> Go to <a
-                        class="govuk-link">@Html.ActionLink("Finalise applications", "FinaliseApplications", "application")</a>
-                    and follow the instructions there. </p>
+                @if (isLA = true)
+                {
+                    <p>
+                        <a href="#" id="print-link" class="govuk-link"> Print this page</a> to give to the parent or guardian as a record of the reference numbers.
+                    </p>
+                }
+                else
+                {
+                    <p>
+                        As these children are entitled to free school meals, you’ll now need to add details of their
+                        application to your own system before finalising.
+                    </p>
+                    <p>
+                        Go to <a class="govuk-link">@Html.ActionLink("Finalise applications", "FinaliseApplications", "application")</a>
+                        and follow the instructions there.
+                    </p>
+                }
             </div>
         </div>
     </div>
-
     <p>@Html.ActionLink("Return to dashboard", "Index", "Home")</p>
 </div>
