+++ CheckYourEligibility.Admin/Views/Shared/_Layout.cshtml
@@ -179,11 +179,10 @@
 
     <script src="~/lib/jquery/dist/jquery.min.js"></script>
     <script src="~/lib/bootstrap/dist/js/bootstrap.bundle.min.js"></script>
-    <script src="~/js/site.js" asp-append-version="true"></script>
     <script type="module" src="~/js/govuk-frontend.min.js"></script>
     <script src="~/js/dfefrontend.js"></script>
 
-    <script type="module" src="~/js/cookieFunctions.js"></script>
+    <script type="module" src="~/js/site.js" asp-append-version="true"></script>
 
     @await RenderSectionAsync("Scripts", false)
 </body>
